changed quantities to be floats and added checking to ensure barcodes are unique

+++ inventory/forms.py
@@ -7,7 +7,7 @@ from inventory.models import Item
 class CreateItemForm(FlaskForm):
     name = StringField(label = 'Name', validators = [Length(min=2, max=30), DataRequired()])
     price = DecimalField(label = 'Price', validators = [DataRequired()])
-    quantity = IntegerField(label = 'Quantity', validators = [DataRequired()])
+    quantity = DecimalField(label = 'Quantity', validators = [DataRequired()])
     barcode = IntegerField(label = 'Barcode', validators = [NumberRange(min = 0, max = 999999999999), DataRequired()])
     description = StringField(label = 'Description', validators = [Length(min=2, max=1024), DataRequired()])
     submitCreating = SubmitField(label = "Create Item")
